+++ CHANGES.md
@@ -15,6 +15,7 @@ Changes in 11.5
 - Use DBUpdate() & DBInsert() functions in Student.php & User.php
 - Fix PHP Fatal error maximum execution time of 120 seconds exceeded in InstallDatabase.php
 - Update French & Spanish translations in rosariosis.po
+- MakeChooseCheckbox() Add 'required' option to $value param. Prevent submitting form if no checkboxes are checked in Inputs.php
 
 Changes in 11.4.4
 -----------------